Moved formatters from 'br/presenter/formatter' to 'br/formatting'.
