+++ brjs-sdk/sdk/libs/javascript/br-formatting/.js-style
@@ -1,0 +1,1 @@
+namespaced-js
