+++ brjs-sdk/sdk/libs/javascript/br-formatting/br-lib.conf
@@ -1,0 +1,1 @@
+requirePrefix: br/formatting
